Rename map-serialize folder to serializers
