+++ src/windshaft/serializers/anonymous-map-serializer/analysis-serializer.js
