+++ src/windshaft/serializers/anonymous-map-serializer/anonymous-map-serializer.js
