+++ src/windshaft/serializers/anonymous-map-serializer/dataviews-serializer.js
