+++ src/windshaft/serializers/anonymous-map-serializer/layers-serializer.js
