+++ src/windshaft/serializers/named-map-serializer/named-map-serializer.js
