+++ test/spec/windshaft/serializers/anonymous-map-serializer/analysis-serializer.spec.js
@@ -3,7 +3,7 @@ var AnalysisService = require('../../../../../src/analysis/analysis-service.js')
 var CartoDBLayer = require('../../../../../src/geo/map/cartodb-layer');
 var VisModel = require('../../../../../src/vis/vis');
 var DataviewModel = require('../../../../../src/dataviews/dataview-model-base');
-var AnalysisSerializer = require('../../../../../src/windshaft/map-serializer/anonymous-map-serializer/analysis-serializer');
+var AnalysisSerializer = require('../../../../../src/windshaft/serializers/anonymous-map-serializer/analysis-serializer');
 
 describe('analysis-serializer', function () {
   var visModel;
